+++ packages/touch_targets/lib/touch_targets.dart
@@ -1,5 +1,8 @@
 library;
 
-export 'src/default_touch_targets.dart';
-export 'src/touch_client.dart';
-export 'src/touch_group.dart';
+export 'src/core/touch_client.dart';
+export 'src/core/touch_group_registry.dart';
+export 'src/core/touch_group.dart';
+export 'src/targets/custom_touch_targets.dart';
+export 'src/targets/default_touch_targets.dart';
+export 'src/targets/touch_target_fit.dart';